Breadcrumb: File rename to make the 6.0 merge diff more readable. (#4394)
* File rename to make the 6.0 merge diff more readable.

* Adding checkbox.

* Also moving a few more.
