+++ packages/office-ui-fabric-react/src/components/Breadcrumb/Breadcrumb.test.tsx
@@ -5,9 +5,7 @@ import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 import * as ReactTestUtils from 'react-dom/test-utils';
 import * as renderer from 'react-test-renderer';
-
-import { Breadcrumb } from './Breadcrumb';
-import { IBreadcrumbItem } from './Breadcrumb.types';
+import { Breadcrumb, IBreadcrumbItem } from './index';
 
 describe('Breadcrumb', () => {
   it('renders breadcumb correctly', () => {
