+++ packages/office-ui-fabric-react/src/components/Breadcrumb/index.ts
@@ -1,2 +1,2 @@
-export * from './Breadcrumb';
+export * from './Breadcrumb.base';
 export * from './Breadcrumb.types';
