+++ packages/office-ui-fabric-react/src/components/Checkbox/Checkbox.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import * as renderer from 'react-test-renderer';
 
-import { Checkbox } from './Checkbox';
+import { Checkbox } from './index';
 
 describe('Checkbox', () => {
   it('renders Checkbox correctly', () => {
