+++ packages/office-ui-fabric-react/src/components/Checkbox/index.ts
@@ -1,2 +1,2 @@
-export * from './Checkbox';
+export * from './Checkbox.base';
 export * from './Checkbox.types';
