+++ packages/office-ui-fabric-react/src/components/Icon/Icon.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import * as renderer from 'react-test-renderer';
 
-import { Icon } from './Icon';
+import { Icon } from './index';
 
 describe('Icon', () => {
   it('renders Icon correctly', () => {
