+++ packages/office-ui-fabric-react/src/components/Label/index.ts
@@ -1,2 +1,2 @@
-export * from './Label';
+export * from './Label.base';
 export * from './Label.types';
